+++ servant-foreign/CHANGELOG.md
@@ -1,4 +1,4 @@
-0.5
+HEAD
 -----
-
+* Use the `text` package instead of `String`.
 * Extract javascript-oblivious types and helpers to *servant-foreign*